+++ NEWS.md
@@ -7,6 +7,7 @@ See the changelog for a full list of changes.
 ----------------------------------
 
 ### Requirements
+- IntelMQ no longer supports Python 3.4, Python `>=` 3.5 is required.
 
 ### Tools
 
@@ -19,7 +20,7 @@ See the changelog for a full list of changes.
 ### Postgres databases
 
 
-2.1.2 Bugfix release (unreleased)
+2.1.3 Bugfix release (unreleased)
 ---------------------------------
 
 ### Requirements
@@ -27,16 +28,25 @@ See the changelog for a full list of changes.
 ### Tools
 
 ### Harmonization
+The regular expression of the field `protocol.transport` has been updated to accommodate the value `nvp-ii`.
+`intelmqctl upgrade-config` handles the change to automatically upgrade your configuration.
 
 ### Configuration
+
+### Libraries
+
+### Postgres databases
+
+
+2.1.2 Bugfix release (2020-01-28)
+---------------------------------
 
 #### MaxMind GeoIP
 MaxMind requires a registration before being able to download the GeoLite2 database starting with 2019-12-30: https://blog.maxmind.com/2019/12/18/significant-changes-to-accessing-and-using-geolite2-databases/
 If the provided `update-geoip-data` script is used, the license key can now be set second parameter.
 
 ### Libraries
-
-### Postgres databases
+When using MISP integrations, make sure your currently installed PyMISP version supports the installed Python version. Any PyMISP version newer than 2.4.119.1 requires Python 3.6 or newer.
 
 
 2.1.1 Bugfix release (2019-11-11)